fix: hostname setting problem for lepton case when cross node replica  (#224)
In lepton for torch run one replica on multiple nodes, we need set
hostname to make the connection work.
But sometimes the previous used `hostname` in lepton manner is too long
and cannot be set.
So, we use the eth ip as the `hostname` to be set, and can work in the
failed cases.
Tested:
grpo:

https://dashboard.dgxc-lepton.nvidia.com/workspace/b5k2m9x7/compute/jobs/detail/ccot-base-lightr1-debug-w97c/replicas/list
sft:

https://dashboard.dgxc-lepton.nvidia.com/workspace/b5k2m9x7/compute/jobs/detail/echo-sft-test-46z4/replicas/list