[Fix] Deepseek V3 GRPO bug fix (#259)
Currently Deepseek V3 runs hit
[this](https://github.com/nvidia-cosmos/cosmos-rl/blob/main/cosmos_rl/utils/parallelism_map.py#L101)
assertion issue when running GRPO when using EP. However, for deepseek
model, given the mix of dense/MoE layers, EP is not supported in
rollout. In other words, rollout assumes that the tensor being
constructed in the full tensor.

This PR removes `deepseek_v3` model from the list of MoE models in
rollout to get around the issue.

As part of the investigation, I also noticed part of the logic in weight
mapper is not used and can be removed to avoid confusion.