feat: Support epoch level save frequency by `save_freq_in_epoch` (#264)
Considering:
- The replica number changing in dynamic scaling which makes the steps
needed for one epoch varying;
- The concurrent control may be needed to save checkpoint at both step
level and epoch level.
So, we don't modify the save frequency in steps as a newly calculated
step interval from the epochs. We rather add a new control besides the
step level to exactly match the epoch level save frequency in the
controller.
Both sft and rl are supported.
`save_freq_in_epoch` in` config.train.ckpt` can be used to specify the
frequency at epoch level for saving checkpoints.