Adjust samples generation for on-policy. (#324)
This PR adjust the samples management of on-policy and DAPO.

For on-policy, We will calculate the samples that needed for one step of
policy. If the samples are not enough, we will dispatch more prompts to
rollout for current step to generate more samples.

Two fields for config:

1. rollout `batch_size`: number of prompts that one rollout replica will
fetch from controller
2. `train.train_policy.max_retry_for_on_policy`: upper limits of rollout
generation of a policy step.

So rollout will consume `batch_size * max_retry_for_on_policy` prompts
to produce enough samples for policy. If still not engough, an error
will emit same as VeRL.

If `max_retry_for_on_policy` is 0 or negative, that means rollout will
consume prompts until samples are enough for current policy step.